Rename component/view, #1136
