+++ afs/tasks.py
@@ -14,7 +14,7 @@ except ImportError:
 
 @shared_task
 def download_related_files(userid, files, project_name):
-    from afs.views.download_files import FileDownloader
+    from afs.views.download_project_files import FileDownloader
 
     downloader = FileDownloader()
     user = User.objects.get(pk=userid)
